Moved to a single examples file
